# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ki-db Knowledge Index API
  version: 1.0.0
  description: |
    Postgres-centric Knowledge Index DB with hybrid BM25+vector search,
    knowledge graph, and event-driven indexing pipeline.
servers:
  - url: http://localhost:8080
    description: Ingest API
  - url: http://localhost:8081
    description: Query API
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
  schemas:
    Document:
      type: object
      required: [doc_id, stable_key, doc_type, status]
      properties:
        tenant_id:
          type: string
          format: uuid
        doc_id:
          type: string
          format: uuid
        stable_key:
          type: string
        title:
          type: string
        doc_type:
          type: string
          enum: [report, adr, postmortem, snippet, glossary, guide, policy, other]
        status:
          type: string
          enum: [inbox, published, deprecated, archived]
        confidence:
          type: string
          enum: [high, med, low]
        owners:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentVersion:
      type: object
      properties:
        version_id:
          type: string
          format: uuid
        version_no:
          type: integer
        content_uri:
          type: string
        content_sha256:
          type: string
        created_at:
          type: string
          format: date-time
    CreateDocumentRequest:
      type: object
      required: [stable_key, doc_type, content]
      properties:
        stable_key:
          type: string
        title:
          type: string
        doc_type:
          type: string
        content:
          type: string
          description: Raw text content of the document
        owners:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - TenantHeader: []
paths:
...
  /api/v1/documents:
    post:
      tags: [Ingest]
      summary: Create or update a document
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [Ingest]
      summary: List documents
      operationId: listDocuments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: doc_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  total:
                    type: integer
  /api/v1/documents/{docID}:
    get:
      tags: [Ingest]
      summary: Get document with versions
      operationId: getDocument
      parameters:
        - name: docID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document with versions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Document'
                  - type: object
                    properties:
                      versions:
                        type: array
                        items:
                          $ref: '#/components/schemas/DocumentVersion'
  /api/v1/documents/{docID}/status:
    patch:
      tags: [Ingest]
      summary: Update document status
      operationId: updateDocumentStatus
      parameters:
        - name: docID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [inbox, published, deprecated, archived]
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  doc_id:
                    type: string
                    format: uuid
                  status:
                    type: string
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Query API ---
...